backend/requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.4
Django==4.1.4
django-cors-headers==3.13.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
jsonfield==3.1.0
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.21.0
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
